Add  JwtUtil , new DTO , SecurityConfig , 2 new migrations , 2 new Enums and rewrite Task and User entity and Repositories ###Description -*Add JWTUtil -*Add LoginDto and RegisterDto -*Add SecurityConfig and JwtUtil -*Add enum UserRole and TaskPriority -*Add migration which add priority and user rile columns -*Rewrite UserRepository and UserService and User entity -*Rewrite Task DTO and TaskService and move in separate package TaskStatus -*Add separate package for enums

+++ src/main/java/org/example/taskmanager/service/TaskService.java
@@ -1,11 +1,15 @@
 package org.example.taskmanager.service;
 
+
 import lombok.RequiredArgsConstructor;
 import org.example.taskmanager.dto.TaskDTO;
 import org.example.taskmanager.entity.Task;
+import org.example.taskmanager.entity.User;
 import org.example.taskmanager.repository.TaskRepository;
 
+import org.example.taskmanager.repository.UserRepository;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -13,20 +17,34 @@ import java.util.List;
 @RequiredArgsConstructor
 public class TaskService {
     private final TaskRepository taskRepository;
-    public Task save(TaskDTO taskDto) {
+    private final UserRepository userRepository;
+
+    @Transactional
+    public Task save(TaskDTO taskDto , String username) {
+        User user  = userRepository.findByUsername(username)
+                .orElseThrow(()-> new RuntimeException("User not founed"));
         Task task = new Task();
         task.setTitle(taskDto.getTitle());
         task.setDescription(taskDto.getDescription());
         task.setStatus(taskDto.getStatus());
+        task.setUser(user);
         return taskRepository.save(task);
     }
+
     public List<Task> findAll() {
         return taskRepository.findAll();
     }
+
     public Task findById(Long id) {
-        return taskRepository.findById(id).orElse(null);
+        return taskRepository.findById(id)
+                .orElseThrow(()-> new RuntimeException("Task not founed"));
     }
+
     public void delete(Long id) {
-        taskRepository.deleteById(id);
+        if(taskRepository.existsById(id)) {
+            taskRepository.deleteById(id);
+        }else{
+            throw new RuntimeException("Task not founed");
+        }
     }
 }
